fix: address code review findings across Gymnasium envs, packaging, and tests
- Wrap observation current_player in np.int64 for Battleship and Checkers to support VectorEnv batching
- Isolate RNG in GravityDuel to prevent global np.random.seed pollution
- Update Pacman step order so power pellet frightened mode applies before ghost collision checks
- Fix equal Box space bounds in Drone env to resolve Gymnasium warnings
- Add check_env test and remove filesystem image side-effects in 2048 env_test.py
- Expose __version__ = "0.0.1" in envpack/__init__.py
- Update MANIFEST.in to include screenshot GIF animations in sdist
- Remove setuptools runtime dependency from requirements.txt
- Remove orphaned game_lunar_transit directory and update .gitignore

+++ envpack/envs/game_battleship/env.py
@@ -204,7 +204,7 @@ class GymBattleshipEnv(gym.Env):
         return {
             "observation": obs_grid,
             "valid_mask": valid_mask,
-            "current_player": self._current_player,
+            "current_player": np.int64(self._current_player),
             "ships_left": np.array([p1_rem, p2_rem], dtype=np.int32),
         }
